+++ application/src/main/java/com/mercadolibre/planning/model/me/gateways/planningmodel/dtos/HeadcountProductivityData.java
@@ -9,5 +9,5 @@ import java.time.OffsetTime;
 @Builder
 public class HeadcountProductivityData {
     private OffsetTime dayTime;
-    private double productivity;
+    private long productivity;
 }
